Add option for allowing the grouping of paths.
For routers that have variables in the path, the HistoGram
instance can become huge. This can be both a memory problem
and a problem for prometheus scrape times.

This change adds an option to group the path in the prom
metrics such that the matching route name is used instead
of the request uri path.